Change bitrate type to long (#245)
* Change bitrate to long

* More bitrate

* And one more test

* More tests

+++ Xabe.FFmpeg/Model/Format.cs
@@ -17,7 +17,7 @@ namespace Xabe.FFmpeg.Model
         {
             public string size { get; set; }
 
-            public double bit_Rate { get; set; }
+            public long bit_Rate { get; set; }
 
             public double duration { get; set; }
         }
